cloud-function/requirements.txt
Flask==2.1.0
google-cloud-firestore==1.8.1